tstern 6 17 2021

+++ test_driver.sh
@@ -1,16 +1,29 @@
 #! /bin/sh
+set -e
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_logs
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs
+fi
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/notebook_testing_dev" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/notebook_testing_dev
+fi
+if [ ! -d "/home/jovyan/opensarlab-notebook_testing/reports" ]; then
+  mkdir /home/jovyan/opensarlab-notebook_testing/reports
+fi
 oldlogs="/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs/*.log"
-for f in $oldlogs
-do
-   rm $f
-done
-# Re-install and activate rtc_analysis conda environment
+if [ -f "$oldlogs" ]; then
+  for f in $oldlogs
+    do
+      rm $f
+    done
+fi
+## Re-install and activate rtc_analysis conda environment
 conda init bash
 source ~/.bashrc
 conda env create -f '/home/jovyan/conda_environments/Environment_Configs/rtc_analysis_env.yml' --prefix "/home/jovyan/.local/envs/rtc_analysis" --force
 conda run -n rtc_analysis kernda --display-name rtc_analysis -o /home/jovyan/.local/envs/rtc_analysis/share/jupyter/kernels/python3/kernel.json
 conda activate /home/jovyan/.local/envs/rtc_analysis
-# Run tests on notebooks that use rtc_analysis environment
+## Run tests on notebooks that use rtc_analysis environment
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex1.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex2.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Ecosystems_Ex3.py
@@ -24,27 +37,43 @@ python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Ex3B.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Ex4A.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Ex4B.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Master_Change_Detection_Amplitude_Time_Series_Example.py
-python /home/jovyan/opensarlab-notebook_testing/Test_ASF_Projects_Prepare_Data_Stack_Hyp3_no_group.py
 python /home/jovyan/opensarlab-notebook_testing/Test_ASF_Projects_Subset_Data_Stack.py
-python /home/jovyan/opensarlab-notebook_testing/Test_Master_Prepare_Data_Stack_Hyp3_no_group.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Master_Subset_Data_Stack.py
-python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Prepare_Data_Stack_Hyp3_no_group.py
 python /home/jovyan/opensarlab-notebook_testing/Test_Hazards_Subset_Data_Stack.py
-# Deactive rtc_analysis environment
+## Deactivate rtc_analysis environment
 conda deactivate
-# Re-install and activate insar_analysis conda environment
+##Re-install and activate insar_analysis conda environment
 conda env create -f '/home/jovyan/conda_environments/Environment_Configs/insar_analysis_env.yml' --prefix "/home/jovyan/.local/envs/insar_analysis" --force
 conda run -n insar_analysis kernda --display-name insar_analysis -o /home/jovyan/.local/envs/insar_analysis/share/jupyter/kernels/python3/kernel.json
-source install_insar_analysis_pkgs.sh
+conda init
+source /home/jovyan/conda_environments/install_insar_analysis_pkgs.sh
 conda activate /home/jovyan/.local/envs/insar_analysis
-# Run tests on notebooks that use insar_analysis environment
+## Run tests on notebooks that use insar_analysis environment
 python /home/jovyan/opensarlab-notebook_testing/Test_Master_InSAR_volcano_source_modeling.py
 python /home/jovyan/opensarlab-notebook_testing/Test_GEOS657_Lab6.py
-# Deactive insar_analysis environment
+## Deactivate insar_analysis environment
+conda deactivate
+##Re-install and activate the machine learning conda environment
+conda env create -f '/home/jovyan/conda_environments/Environment_Configs/machine_learning_env.yml' --prefix "/home/jovyan/.local/envs/machine_learning" --force
+conda run -n machine_learning kernda --display-name machine_learning -o /home/jovyan/.local/envs/machine_learning/share/jupyter/kernels/python3/kernel.json
+conda init
+conda activate /home/jovyan/.local/envs/machine_learning
+## Run tests on notebooks that use the machine_learning environment
+python /home/jovyan/opensarlab-notebook_testing/Test_Master_CRNN_change_detection.py
+## Deactivate machine_learning environment
+conda deactivate
+##Re-install and activate the train conda environment
+conda env create -f '/home/jovyan/conda_environments/Environment_Configs/train_env.yml' --prefix "/home/jovyan/.local/envs/train" --force
+conda run -n train kernda --display-name train -o /home/jovyan/.local/envs/train/share/jupyter/kernels/python3/kernel.json
+conda init
+conda activate /home/jovyan/.local/envs/train
+## Run tests on notebooks that use the train environment
+python /home/jovyan/opensarlab-notebook_testing/Test_GEOS_657_2019_Lab9_InSARTimeSeriesAnalysis_Part1_DataDownload_HyP3_v2.py
+## Deactivate train environment
 conda deactivate
 echo 'Done running tests on the notebooks!'
 echo 'Check the logs for exceptions and failures!'
-# Get rid of spaces in log filenames
+## Get rid of spaces in log filenames
 oldnames="/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/*.log"
 for f in $oldnames
 do
@@ -90,6 +119,6 @@ for f in $infologs
 do
    mv $f "/home/jovyan/opensarlab-notebook_testing/notebook_testing_logs/old_logs/"
 done
-#mail -s "OSL Notebook Test Report" testern@alaska.edu <  test_report_$now.txt
+##mail -s "OSL Notebook Test Report" testern@alaska.edu <  test_report_$now.txt
 mv test_report_$now.txt "/home/jovyan/opensarlab-notebook_testing/reports/"
 echo 'Issues: '$issue_counter
